Refactor dropdown logic
