+++ Fitlance/ClientApp/src/components/findTrainers/hooks/useDropdownControl.js
@@ -1,0 +1,23 @@
+import { useState, useEffect, useRef } from 'react'; 
+
+const useDropdownControl = () => {
+    const [isOpen, setIsOpen] = useState(false);
+    const dropdownRef = useRef(null);
+
+    const handleClickOutside = (event) => {
+        if (dropdownRef.current && !dropdownRef.current.contains(event.target)) {
+            setIsOpen(false);
+        }
+    };
+
+    useEffect(() => {
+        document.addEventListener("mousedown", handleClickOutside);
+        return () => {
+            document.removeEventListener("mousedown", handleClickOutside);
+        };
+    }, []);
+
+    return { isOpen, setIsOpen, dropdownRef }
+};
+
+export default useDropdownControl;
